+++ raycast-env.d.ts
@@ -15,10 +15,14 @@ declare type Preferences = ExtensionPreferences
 declare namespace Preferences {
   /** Preferences accessible in the `manage-replacements` command */
   export type ManageReplacements = ExtensionPreferences & {}
+  /** Preferences accessible in the `create-replacement` command */
+  export type CreateReplacement = ExtensionPreferences & {}
 }
 
 declare namespace Arguments {
   /** Arguments passed to the `manage-replacements` command */
   export type ManageReplacements = {}
+  /** Arguments passed to the `create-replacement` command */
+  export type CreateReplacement = {}
 }